Move stacking test to test_shell
